requirements.txt
Django>=5.0,<6.0
psycopg2-binary>=2.9,<3.0
python-dotenv>=1.0,<2.0
django-encrypted-model-fields>=0.6,<1.0
Pillow>=10.0,<11.0
Markdown>=3.5,<4.0
weaviate-client>=4.0,<5.0
openai>=1.0,<2.0
google-generativeai>=0.3,<1.0
